Fix PAM

+++ Core/Inc/PAM/pam_config.h
@@ -48,8 +48,8 @@ typedef struct{
 
 	I2C_TypeDef *I2Cx_PORT;
 	uint8_t I2C_addr_GPIO_Ext;
-	uint16_t pin_State_ID;
-	uint16_t pin_Enable_eF;
+	uint32_t pin_State_ID;
+	uint32_t pin_Enable_eF;
 
 	uint8_t I2C_addr_PWR_Mon;
 	uint16_t PWR_Mon_Max_Current_int16;
